intra-molecular interactions for the solute weren't being included correctly.
Previously, we turned off all interactions of the alchemical group in the customNonbondedForce: ""
for index in self.solute_total_atoms_list:
    self.customNonbondedForce.setParticleParameters(index, [ 0.0 , 0.0 , 0.0 , 0.0 , 10.0 , 0.0 , 0.0 , 0.0 , 0.0 ] )
""
furthermore, in the customNonbondedForce_alchemical class, we only included interactions between alchemical group, and non-alchemical group, e.g. ""
self.customNonbondedForce_alchemical[force_index].addInteractionGroup( self.alchemical_local , self.nonalchemical_atomset )
""
so this means that intra-molecular interactions of the alchemical group were completely omitted. Obviously, we
need to keep intra-molecular interactions.

the best way to do this is to keep the intra-molecular interactions of the alchemical group in the original customNonbondedForce class, by not
zeroing parameters, but setting up exclusions instead---this is what we do in the new, modified version.
Note that the alternative of including intra-molecular interactions in the new customNonbondedForce_alchemical force class is messy, because when
we make multiple force instances for repulsion (dividing up the molecule) it is not easy to avoid multiple counting of the same intra-molecular interaction.

Note that all previous simulations were unaffected by this, as BF4, PF6, NO3, triflate, do not have intra-molecular interactions, and the only intermolecular interactions within TFSI are included as
a CustomBondForce force type, which are not modified during thermodynamic integration