Fixed a bug in the thermodynamic integration of non-bonded interactions.  Before, in the TI class, we had the following code
          # don't zero parameters as in commented above code, because we want to keep intra-molecular alchemical interactions within this force class ...
          for i_index in self.alchemical_atomset:
              for j_index in self.nonalchemical_atomset:
                  self.customNonbondedForce.addExclusion(i_index,j_index)
                  if interaction_type == "repulsion":
                      # need to add exceptions for electrostatics, even though charges are zero, have numerical problems in electrostatics when repulsion is scaled down
                      self.nbondedForce.addException(i_index,j_index,0,1,0,True)

 This code only excluded electrostatics for the alchemical group for interaction type "repulsion", as we assumed that charges had been turned off for other non-bonded interaction types (e.g. VDW or SAPT_FF_LJ scale).  However, this wasn't true.  Because we create a new simulation object for each interaction_type that we are turning off, even though charges are scaled to zero during the electrostatic TI, they are turned back to their original values in the subsequent interaction scalings when a new simulation object is created.

This bug has been fixed.  In addition, we have introduced soft-core potential scaling for non-bonded interactions.  First, we scale the SAPT-FF potential to a LJ-like potential (LJ plus C8/R^8).  C12 repulsive parameter is chosen such that it gives 1 kJ/mol repulsion at the same distance as the Born-Mayer potential of SAPT-FF.  After this, we then scale off the soft-core LJ interaction using a similar functional form as given in the Gromacs Manual.

We have also made the code versatile to either do TI on a SAPT-FF force field, or normal LJ forcefield (which we use to benchmark OPLS Cl- ion in tip4pew water)